Fix use of tqdm to avoid creating many unnecessary TMonitor threads
Previously a tqdm subclass was being defined as a local subclass
inside a function, which created a new class object for every
invocation of the function. tqdm creates a new TMonitor thread
for every one of its subclasses, so a separate TMonitor thread
was created for each call of the above-mentioned function.

Fix the problem by creating the tqdm subclass outside the function
to only create the class object once.