Close all database Cursors explicitly
This was hoped to fix a case where a final database write
wasn't actually committed to the database before it closed,
but a different change was the real fix.

Nevertheless I think the explicit scoping of Cursor
operations makes the code easier to reason about,
so I'm leaving the scopes in.