run_tests_in_parallel.py: Runs 2 Crystal --test subprocesses in parallel
The tests to run are chosen up-front.

There is no redistribution of load. Thus one of the subprocesses can become
idle if the other one is given more expensive tests to run.

The output format is the same as the format that `crystal --test` uses,
with "[" lines like "[Runner] ..." and "[Worker 0] ..." interleaved to
show status updates from the parallel runner system.